+++ CHANGELOG.md
@@ -5,9 +5,11 @@ IMPROVEMENTS
 * [\#53](https://github.com/binance-chain/go-sdk/pull/53) [SOURCE] change the default source into 0
 * [\#56](https://github.com/binance-chain/go-sdk/pull/56) [RPC] add reconnect strategy when timeout to receive response
 * [\#61](https://github.com/binance-chain/go-sdk/pull/61) [KEY] support bip44 to derive many address from same seed phase
+* [\#71](https://github.com/binance-chain/go-sdk/pull/71) [RPC] add timelock query support 
 
 FEATURES
-*[\#66](https://github.com/binance-chain/go-sdk/pull/66)  [API]  support set account flag transaction
+
+* [\#66](https://github.com/binance-chain/go-sdk/pull/66)  [API]  support set account flag transaction
 
 BREAKING
 * [\#57](https://github.com/binance-chain/go-sdk/pull/57) [API] add query option to getTokens api